+++ converter/firmware/tk_resolution.h
@@ -1,0 +1,7 @@
+
+#include "tk_input_converter.h"
+
+
+void reso_calo(pt_t pt, etaphi_t eta_calo, pt_t& err);
+
+
