+++ converter/run.tcl
@@ -7,7 +7,10 @@ set topFunc ${mainAlgo}_hw
 # open the project, don't forget to reset
 open_project -reset "track-conv-test"
 set_top ${topFunc}
-add_files firmware/wrapper.cpp 
+#add_files firmware/tk_pt_inversion.cpp
+add_files firmware/tk_input_converter.cpp 
+add_files firmware/tk_prop_tanlambda.cpp
+add_files firmware/tk_resolution.cpp
 add_files -tb test.cpp
 
 # reset the solution
@@ -19,9 +22,9 @@ set_clock_uncertainty 1.5
 config_interface -trim_dangling_port
 # do stuff
 csim_design
-csynth_design
-cosim_design -trace_level all
-export_design -format ip_catalog -vendor "cern-cms" -version 0.1 -description "${mainAlgo}"
+#csynth_design
+# cosim_design -trace_level all
+# export_design -format ip_catalog -vendor "cern-cms" -version 0.1 -description "${mainAlgo}"
 
 # exit Vivado HLS
 exit
